requirements.txt
pandas
--extra-index-url https://download.pytorch.org/whl/cu113
torch
--extra-index-url https://download.pytorch.org/whl/cu113
torchvision
--extra-index-url https://download.pytorch.org/whl/cu113
torchaudio
rdkit-pypi
dgllife
pytorch-lightning
-f https://data.dgl.ai/wheels/repo.html
dgl-cu113
-f https://data.dgl.ai/wheels/repo.html
dglgo